refactor(#2631): Phase 3 — unify persistence layer between conductor-core and runkon-flow
Deletes the duplicate WorkflowPersistence trait, NewRun/NewStep/StepUpdate
structs, FanOutItemRow, and in-memory impl from conductor-core. All
consumers now use runkon_flow::traits::persistence directly.

Also adds utoipa support to runkon-flow's FanOutItemRow and ports the
impl Into<String> ergonomics improvement to StepUpdate::failed.

Impact: ~1,617 lines deleted, zero functional change.

Closes #2631 (Phase 3)
