+++ conductor-core/src/workflow/executors/resolvers/human_approval.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use crate::error::{ConductorError, Result};
 use crate::workflow::executors::gate_resolver::{GateContext, GateParams, GatePoll, GateResolver};
-use crate::workflow::persistence::{GateApprovalState, WorkflowPersistence};
+use runkon_flow::traits::persistence::{GateApprovalState, WorkflowPersistence};
 
 /// Distinguishes the two human gate types so a single struct can register
 /// under both `"human_approval"` and `"human_review"`.
@@ -57,8 +57,8 @@ impl GateResolver for HumanApprovalGateResolver {
 mod tests {
     use super::*;
     use crate::workflow::executors::gate_resolver::{GateContext, GateParams};
-    use crate::workflow::persistence::WorkflowPersistence;
     use crate::workflow::persistence_sqlite::SqliteWorkflowPersistence;
+    use runkon_flow::traits::persistence::WorkflowPersistence;
     use runkon_flow::dsl::ApprovalMode;
     use rusqlite::Connection;
     use std::sync::Arc;
