+++ conductor-core/src/workflow/runkon_bridge.rs
@@ -14,120 +14,6 @@ use crate::error::ConductorError;
 use crate::workflow::action_executor::ActionExecutor;
 use crate::workflow::item_provider::ItemProvider;
 
-// ---------------------------------------------------------------------------
-// Persistence type conversions (runkon-flow ↔ conductor-core)
-// ---------------------------------------------------------------------------
-// These remain necessary until Phase 3 unifies the persistence traits.
-
-macro_rules! field_copy_from {
-    ($from:ty => $to:ty { $($field:ident),+ $(,)? }) => {
-        impl From<$from> for $to {
-            fn from(src: $from) -> Self {
-                Self {
-                    $($field: src.$field),+
-                }
-            }
-        }
-    };
-}
-
-field_copy_from!(runkon_flow::traits::persistence::NewRun => crate::workflow::persistence::NewRun {
-    workflow_name,
-    worktree_id,
-    ticket_id,
-    repo_id,
-    parent_run_id,
-    dry_run,
-    trigger,
-    definition_snapshot,
-    parent_workflow_run_id,
-    target_label,
-});
-
-field_copy_from!(runkon_flow::traits::persistence::NewStep => crate::workflow::persistence::NewStep {
-    workflow_run_id,
-    step_name,
-    role,
-    can_commit,
-    position,
-    iteration,
-    retry_count,
-});
-
-field_copy_from!(runkon_flow::traits::persistence::StepUpdate => crate::workflow::persistence::StepUpdate {
-    status,
-    child_run_id,
-    result_text,
-    context_out,
-    markers_out,
-    retry_count,
-    structured_output,
-    step_error,
-});
-
-field_copy_from!(crate::workflow::persistence::FanOutItemRow => runkon_flow::types::FanOutItemRow {
-    id,
-    step_run_id,
-    item_type,
-    item_id,
-    item_ref,
-    child_run_id,
-    status,
-    dispatched_at,
-    completed_at,
-});
-
-impl From<runkon_flow::traits::persistence::FanOutItemStatus> for crate::workflow::persistence::FanOutItemStatus {
-    fn from(s: runkon_flow::traits::persistence::FanOutItemStatus) -> Self {
-        match s {
-            runkon_flow::traits::persistence::FanOutItemStatus::Pending => Self::Pending,
-            runkon_flow::traits::persistence::FanOutItemStatus::Running => Self::Running,
-            runkon_flow::traits::persistence::FanOutItemStatus::Completed => Self::Completed,
-            runkon_flow::traits::persistence::FanOutItemStatus::Failed => Self::Failed,
-            runkon_flow::traits::persistence::FanOutItemStatus::Skipped => Self::Skipped,
-        }
-    }
-}
-
-impl From<crate::workflow::persistence::FanOutItemStatus> for runkon_flow::traits::persistence::FanOutItemStatus {
-    fn from(s: crate::workflow::persistence::FanOutItemStatus) -> Self {
-        match s {
-            crate::workflow::persistence::FanOutItemStatus::Pending => Self::Pending,
-            crate::workflow::persistence::FanOutItemStatus::Running => Self::Running,
-            crate::workflow::persistence::FanOutItemStatus::Completed => Self::Completed,
-            crate::workflow::persistence::FanOutItemStatus::Failed => Self::Failed,
-            crate::workflow::persistence::FanOutItemStatus::Skipped => Self::Skipped,
-        }
-    }
-}
-
-impl From<runkon_flow::traits::persistence::FanOutItemUpdate> for crate::workflow::persistence::FanOutItemUpdate {
-    fn from(u: runkon_flow::traits::persistence::FanOutItemUpdate) -> Self {
-        match u {
-            runkon_flow::traits::persistence::FanOutItemUpdate::Running { child_run_id } => {
-                Self::Running { child_run_id }
-            }
-            runkon_flow::traits::persistence::FanOutItemUpdate::Terminal { status } => {
-                Self::Terminal { status: status.into() }
-            }
-        }
-    }
-}
-
-impl From<crate::workflow::persistence::GateApprovalState> for runkon_flow::traits::persistence::GateApprovalState {
-    fn from(s: crate::workflow::persistence::GateApprovalState) -> Self {
-        match s {
-            crate::workflow::persistence::GateApprovalState::Pending => Self::Pending,
-            crate::workflow::persistence::GateApprovalState::Approved { feedback, selections } => {
-                Self::Approved { feedback, selections }
-            }
-            crate::workflow::persistence::GateApprovalState::Rejected { feedback } => {
-                Self::Rejected { feedback }
-            }
-        }
-    }
-}
-
 /// Convert `ConductorError` to `EngineError`, preserving the cancellation
 /// signal: `WorkflowCancelled` maps to `Cancelled`, all other errors to
 /// `Workflow`.  Centralising this avoids the special-case match being
@@ -607,170 +493,7 @@ impl runkon_flow::engine::ChildWorkflowRunner for ConductorChildWorkflowRunner {
 }
 
 // ---------------------------------------------------------------------------
-// 6. PersistenceAdapter
-// ---------------------------------------------------------------------------
-
-/// Bridges the conductor-core `WorkflowPersistence` trait to the runkon-flow
-/// `WorkflowPersistence` trait, performing type conversion at the boundary.
-///
-/// This is the single place in conductor-core that implements
-/// `runkon_flow::traits::persistence::WorkflowPersistence`; all other modules
-/// interact with the core trait only.
-pub(super) struct PersistenceAdapter(
-    pub(super) Arc<dyn crate::workflow::persistence::WorkflowPersistence>,
-);
-
-impl runkon_flow::traits::persistence::WorkflowPersistence for PersistenceAdapter {
-    fn create_run(
-        &self,
-        new_run: runkon_flow::traits::persistence::NewRun,
-    ) -> Result<runkon_flow::types::WorkflowRun, EngineError> {
-        let core_run = self.0.create_run(new_run.into())?;
-        Ok(core_run)
-    }
-
-    fn get_run(
-        &self,
-        run_id: &str,
-    ) -> Result<Option<runkon_flow::types::WorkflowRun>, EngineError> {
-        let result = self.0.get_run(run_id)?;
-        Ok(result)
-    }
-
-    fn list_active_runs(
-        &self,
-        statuses: &[runkon_flow::status::WorkflowRunStatus],
-    ) -> Result<Vec<runkon_flow::types::WorkflowRun>, EngineError> {
-        let result = self.0.list_active_runs(statuses)?;
-        Ok(result)
-    }
-
-    fn update_run_status(
-        &self,
-        run_id: &str,
-        status: runkon_flow::status::WorkflowRunStatus,
-        result_summary: Option<&str>,
-        error: Option<&str>,
-    ) -> Result<(), EngineError> {
-        self.0
-            .update_run_status(run_id, status.into(), result_summary, error)
-    }
-
-    fn insert_step(
-        &self,
-        new_step: runkon_flow::traits::persistence::NewStep,
-    ) -> Result<String, EngineError> {
-        self.0.insert_step(new_step.into())
-    }
-
-    fn update_step(
-        &self,
-        step_id: &str,
-        update: runkon_flow::traits::persistence::StepUpdate,
-    ) -> Result<(), EngineError> {
-        self.0.update_step(step_id, update.into())
-    }
-
-    fn get_steps(
-        &self,
-        run_id: &str,
-    ) -> Result<Vec<runkon_flow::types::WorkflowRunStep>, EngineError> {
-        let result = self.0.get_steps(run_id)?;
-        Ok(result)
-    }
-
-    fn insert_fan_out_item(
-        &self,
-        step_run_id: &str,
-        item_type: &str,
-        item_id: &str,
-        item_ref: &str,
-    ) -> Result<String, EngineError> {
-        self.0
-            .insert_fan_out_item(step_run_id, item_type, item_id, item_ref)
-    }
-
-    fn update_fan_out_item(
-        &self,
-        item_id: &str,
-        update: runkon_flow::traits::persistence::FanOutItemUpdate,
-    ) -> Result<(), EngineError> {
-        self.0.update_fan_out_item(item_id, update.into())
-    }
-
-    fn get_fan_out_items(
-        &self,
-        step_run_id: &str,
-        status_filter: Option<runkon_flow::traits::persistence::FanOutItemStatus>,
-    ) -> Result<Vec<runkon_flow::types::FanOutItemRow>, EngineError> {
-        let core_filter = status_filter.map(Into::into);
-        let result = self.0.get_fan_out_items(step_run_id, core_filter)?;
-        Ok(result.into_iter().map(Into::into).collect())
-    }
-
-    fn get_gate_approval(
-        &self,
-        step_id: &str,
-    ) -> Result<runkon_flow::traits::persistence::GateApprovalState, EngineError> {
-        let result = self.0.get_gate_approval(step_id)?;
-        Ok(result.into())
-    }
-
-    fn approve_gate(
-        &self,
-        step_id: &str,
-        approved_by: &str,
-        feedback: Option<&str>,
-        selections: Option<&[String]>,
-    ) -> Result<(), EngineError> {
-        self.0
-            .approve_gate(step_id, approved_by, feedback, selections)
-    }
-
-    fn reject_gate(
-        &self,
-        step_id: &str,
-        rejected_by: &str,
-        feedback: Option<&str>,
-    ) -> Result<(), EngineError> {
-        self.0.reject_gate(step_id, rejected_by, feedback)
-    }
-
-    fn is_run_cancelled(&self, run_id: &str) -> Result<bool, EngineError> {
-        self.0.is_run_cancelled(run_id)
-    }
-
-    fn tick_heartbeat(&self, run_id: &str) -> Result<(), EngineError> {
-        self.0.tick_heartbeat(run_id)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn persist_metrics(
-        &self,
-        run_id: &str,
-        input_tokens: i64,
-        output_tokens: i64,
-        cache_read_input_tokens: i64,
-        cache_creation_input_tokens: i64,
-        cost_usd: f64,
-        num_turns: i64,
-        duration_ms: i64,
-    ) -> Result<(), EngineError> {
-        self.0.persist_metrics(
-            run_id,
-            input_tokens,
-            output_tokens,
-            cache_read_input_tokens,
-            cache_creation_input_tokens,
-            cost_usd,
-            num_turns,
-            duration_ms,
-        )
-    }
-}
-
-// ---------------------------------------------------------------------------
-// 7. Helper builder functions
+// 6. Helper builder functions
 // ---------------------------------------------------------------------------
 
 /// Build a runkon-flow `ActionRegistry` backed by a `RkActionExecutorAdapter`
@@ -834,193 +557,6 @@ pub(super) fn build_rk_script_env_provider(
 mod tests {
     use super::*;
     use std::collections::HashMap;
-
-    // ---------------------------------------------------------------------------
-    // Persistence type From impls
-    // ---------------------------------------------------------------------------
-
-    #[test]
-    fn new_run_from_rk_maps_all_fields() {
-        let rk = runkon_flow::traits::persistence::NewRun {
-            workflow_name: "wf".into(),
-            worktree_id: Some("wt".into()),
-            ticket_id: Some("t1".into()),
-            repo_id: Some("r1".into()),
-            parent_run_id: "parent".into(),
-            dry_run: true,
-            trigger: "hook".into(),
-            definition_snapshot: Some("snap".into()),
-            parent_workflow_run_id: Some("pwf".into()),
-            target_label: Some("label".into()),
-        };
-        let core: crate::workflow::persistence::NewRun = rk.into();
-        assert_eq!(core.workflow_name, "wf");
-        assert_eq!(core.worktree_id, Some("wt".into()));
-        assert_eq!(core.ticket_id, Some("t1".into()));
-        assert_eq!(core.repo_id, Some("r1".into()));
-        assert_eq!(core.parent_run_id, "parent");
-        assert!(core.dry_run);
-        assert_eq!(core.trigger, "hook");
-        assert_eq!(core.definition_snapshot, Some("snap".into()));
-        assert_eq!(core.parent_workflow_run_id, Some("pwf".into()));
-        assert_eq!(core.target_label, Some("label".into()));
-    }
-
-    #[test]
-    fn new_step_from_rk_maps_all_fields() {
-        let rk = runkon_flow::traits::persistence::NewStep {
-            workflow_run_id: "run-1".into(),
-            step_name: "step-a".into(),
-            role: "agent".into(),
-            can_commit: true,
-            position: 3,
-            iteration: 2,
-            retry_count: Some(1),
-        };
-        let core: crate::workflow::persistence::NewStep = rk.into();
-        assert_eq!(core.workflow_run_id, "run-1");
-        assert_eq!(core.step_name, "step-a");
-        assert_eq!(core.role, "agent");
-        assert!(core.can_commit);
-        assert_eq!(core.position, 3);
-        assert_eq!(core.iteration, 2);
-        assert_eq!(core.retry_count, Some(1));
-    }
-
-    #[test]
-    fn step_update_from_rk_maps_all_fields() {
-        let rk = runkon_flow::traits::persistence::StepUpdate {
-            status: runkon_flow::status::WorkflowStepStatus::Completed,
-            child_run_id: Some("child".into()),
-            result_text: Some("ok".into()),
-            context_out: Some("ctx".into()),
-            markers_out: Some("mk".into()),
-            retry_count: Some(2),
-            structured_output: Some("{}".into()),
-            step_error: None,
-        };
-        let core: crate::workflow::persistence::StepUpdate = rk.into();
-        assert_eq!(core.status, crate::workflow::WorkflowStepStatus::Completed);
-        assert_eq!(core.child_run_id, Some("child".into()));
-        assert_eq!(core.result_text, Some("ok".into()));
-        assert_eq!(core.context_out, Some("ctx".into()));
-        assert_eq!(core.markers_out, Some("mk".into()));
-        assert_eq!(core.retry_count, Some(2));
-        assert_eq!(core.structured_output, Some("{}".into()));
-        assert!(core.step_error.is_none());
-    }
-
-    #[test]
-    fn fan_out_item_status_from_rk_maps_all_variants() {
-        use runkon_flow::traits::persistence::FanOutItemStatus as Rk;
-        use crate::workflow::persistence::FanOutItemStatus as Core;
-        assert!(matches!(Core::from(Rk::Pending), Core::Pending));
-        assert!(matches!(Core::from(Rk::Running), Core::Running));
-        assert!(matches!(Core::from(Rk::Completed), Core::Completed));
-        assert!(matches!(Core::from(Rk::Failed), Core::Failed));
-        assert!(matches!(Core::from(Rk::Skipped), Core::Skipped));
-    }
-
-    #[test]
-    fn fan_out_item_status_to_rk_maps_all_variants() {
-        use runkon_flow::traits::persistence::FanOutItemStatus as Rk;
-        use crate::workflow::persistence::FanOutItemStatus as Core;
-        assert!(matches!(Rk::from(Core::Pending), Rk::Pending));
-        assert!(matches!(Rk::from(Core::Running), Rk::Running));
-        assert!(matches!(Rk::from(Core::Completed), Rk::Completed));
-        assert!(matches!(Rk::from(Core::Failed), Rk::Failed));
-        assert!(matches!(Rk::from(Core::Skipped), Rk::Skipped));
-    }
-
-    #[test]
-    fn fan_out_item_update_from_rk_running_maps_child_run_id() {
-        let rk = runkon_flow::traits::persistence::FanOutItemUpdate::Running {
-            child_run_id: "child-1".into(),
-        };
-        let core: crate::workflow::persistence::FanOutItemUpdate = rk.into();
-        match core {
-            crate::workflow::persistence::FanOutItemUpdate::Running { child_run_id } => {
-                assert_eq!(child_run_id, "child-1");
-            }
-            other => panic!("expected Running variant, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn fan_out_item_update_from_rk_terminal_maps_status() {
-        let rk = runkon_flow::traits::persistence::FanOutItemUpdate::Terminal {
-            status: runkon_flow::traits::persistence::FanOutItemStatus::Completed,
-        };
-        let core: crate::workflow::persistence::FanOutItemUpdate = rk.into();
-        match core {
-            crate::workflow::persistence::FanOutItemUpdate::Terminal { status } => {
-                assert_eq!(status, crate::workflow::persistence::FanOutItemStatus::Completed);
-            }
-            other => panic!("expected Terminal variant, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn gate_approval_state_to_rk_pending() {
-        let core = crate::workflow::persistence::GateApprovalState::Pending;
-        let rk: runkon_flow::traits::persistence::GateApprovalState = core.into();
-        assert!(matches!(rk, runkon_flow::traits::persistence::GateApprovalState::Pending));
-    }
-
-    #[test]
-    fn gate_approval_state_to_rk_approved() {
-        let core = crate::workflow::persistence::GateApprovalState::Approved {
-            feedback: Some("lgtm".into()),
-            selections: Some(vec!["a".into()]),
-        };
-        let rk: runkon_flow::traits::persistence::GateApprovalState = core.into();
-        match rk {
-            runkon_flow::traits::persistence::GateApprovalState::Approved { feedback, selections } => {
-                assert_eq!(feedback, Some("lgtm".into()));
-                assert_eq!(selections, Some(vec!["a".into()]));
-            }
-            other => panic!("expected Approved, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn gate_approval_state_to_rk_rejected() {
-        let core = crate::workflow::persistence::GateApprovalState::Rejected {
-            feedback: Some("nope".into()),
-        };
-        let rk: runkon_flow::traits::persistence::GateApprovalState = core.into();
-        match rk {
-            runkon_flow::traits::persistence::GateApprovalState::Rejected { feedback } => {
-                assert_eq!(feedback, Some("nope".into()));
-            }
-            other => panic!("expected Rejected, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn fan_out_item_row_from_core_maps_all_fields() {
-        let core = crate::workflow::persistence::FanOutItemRow {
-            id: "id-1".into(),
-            step_run_id: "step-1".into(),
-            item_type: "ticket".into(),
-            item_id: "t-1".into(),
-            item_ref: "ref-1".into(),
-            child_run_id: Some("child-1".into()),
-            status: "completed".into(),
-            dispatched_at: Some("2024-01-01".into()),
-            completed_at: Some("2024-01-02".into()),
-        };
-        let rk: runkon_flow::types::FanOutItemRow = core.into();
-        assert_eq!(rk.id, "id-1");
-        assert_eq!(rk.step_run_id, "step-1");
-        assert_eq!(rk.item_type, "ticket");
-        assert_eq!(rk.item_id, "t-1");
-        assert_eq!(rk.item_ref, "ref-1");
-        assert_eq!(rk.child_run_id, Some("child-1".into()));
-        assert_eq!(rk.status, "completed");
-        assert_eq!(rk.dispatched_at, Some("2024-01-01".into()));
-        assert_eq!(rk.completed_at, Some("2024-01-02".into()));
-    }
 
     // ---------------------------------------------------------------------------
     // delegate_items — mutex poison propagates as EngineError
@@ -1089,70 +625,7 @@ mod tests {
     }
 
     // ---------------------------------------------------------------------------
-    // Fix 9: PersistenceAdapter::get_gate_approval via the adapter
-    // ---------------------------------------------------------------------------
-
-    #[test]
-    fn persistence_adapter_get_gate_approval_returns_approved_state() {
-        use crate::workflow::persistence::{NewRun, NewStep, WorkflowPersistence};
-        use crate::workflow::persistence_memory::InMemoryWorkflowPersistence;
-        use runkon_flow::traits::persistence::GateApprovalState as RkGateApprovalState;
-        use runkon_flow::traits::persistence::WorkflowPersistence as RkPersistence;
-
-        // 1. Create the in-memory persistence.
-        let persistence = InMemoryWorkflowPersistence::new();
-
-        // 2. Create a run and insert a step.
-        let run = persistence
-            .create_run(NewRun {
-                workflow_name: "test-wf".to_string(),
-                worktree_id: None,
-                ticket_id: None,
-                repo_id: None,
-                parent_run_id: "parent".to_string(),
-                dry_run: false,
-                trigger: "test".to_string(),
-                definition_snapshot: None,
-                parent_workflow_run_id: None,
-                target_label: None,
-            })
-            .unwrap();
-        let step_id = persistence
-            .insert_step(NewStep {
-                workflow_run_id: run.id.clone(),
-                step_name: "approval-gate".to_string(),
-                role: "gate".to_string(),
-                can_commit: false,
-                position: 0,
-                iteration: 0,
-                retry_count: None,
-            })
-            .unwrap();
-
-        // 3. Approve the gate via the core persistence.
-        persistence
-            .approve_gate(&step_id, "tester", Some("lgtm"), None)
-            .unwrap();
-
-        // 4. Wrap in PersistenceAdapter.
-        let adapter = PersistenceAdapter(Arc::new(persistence));
-
-        // 5. Call get_gate_approval via the runkon-flow trait and verify the result.
-        let state = RkPersistence::get_gate_approval(&adapter, &step_id).unwrap();
-        match state {
-            RkGateApprovalState::Approved { feedback, .. } => {
-                assert_eq!(
-                    feedback.as_deref(),
-                    Some("lgtm"),
-                    "feedback must survive the approve_gate/get_gate_approval roundtrip via PersistenceAdapter"
-                );
-            }
-            other => panic!("expected Approved, got {other:?}"),
-        }
-    }
-
-    // ---------------------------------------------------------------------------
-    // Fix 10: dependencies method via a concrete Rk item-provider adapter
+    // Fix 9: dependencies method via a concrete Rk item-provider adapter
     // ---------------------------------------------------------------------------
 
     #[test]
